Split out GSL tests from base ones
Move GSL tests into Numeric.GSL.Tests, separate the main into TestBase.hs and
TestGSL.hs.

In hmatrix-tests.cabal:

- Split the test suite into a -base and -gsl ones
- Add a `gsl` configuration flag to select GSL tests
- Add a benchmark section

One can now run hmatrix-base tests suite and benchmarks with:

    cabal configure --flag=-gsl --enable-tests --enable-benchmarks
    cabal tests
    cabal bench
