+++ packages/tests/src/Benchmark.hs
@@ -1,0 +1,3 @@
+import Numeric.LinearAlgebra.Tests
+
+main = runBenchmarks
