+++ packages/tests/src/TestGSL.hs
@@ -1,0 +1,3 @@
+import Numeric.GSL.Tests
+
+main = runTests 20
